A. Schema/Migration Alignment: Fix drift and enforce production-safe types
Changes:
1. Migration: Add missing unique constraint uq_runs_tenant_idempotency
   - Cleaned duplicate test data before applying constraint
   - Migration applied successfully: b705342a947d

2. Models.py: Integer → BIGINT for autoincrement IDs
   - tenant_plans.id: Integer → BIGINT (production scale)
   - tenant_usage_daily.id: Integer → BIGINT (production scale)
   - Removed Integer import (no longer used)

Verification:
- alembic check: "No new upgrade operations detected" ✅
- DB constraint confirmed: uq_runs_tenant_idempotency on (tenant_id, idempotency_key) ✅
- Idempotency tests PASSED (2/2) ✅
- No schema drift detected ✅

P0-A complete ✅

+++ dpp/apps/api/dpp_api/db/models.py
@@ -3,7 +3,7 @@
 from datetime import date, datetime, timezone
 from typing import Optional
 
-from sqlalchemy import BIGINT, DATE, FLOAT, JSON, TEXT, TIMESTAMP, UUID, Index, Integer, UniqueConstraint
+from sqlalchemy import BIGINT, DATE, FLOAT, JSON, TEXT, TIMESTAMP, UUID, Index, UniqueConstraint
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -189,7 +189,8 @@ class TenantPlan(Base):
 
     __tablename__ = "tenant_plans"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    # P0-A: Use BIGINT for autoincrement IDs (production scale)
+    id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
     tenant_id: Mapped[str] = mapped_column(TEXT, nullable=False, index=True)
     plan_id: Mapped[str] = mapped_column(TEXT, nullable=False)
 
@@ -229,7 +230,8 @@ class TenantUsageDaily(Base):
 
     __tablename__ = "tenant_usage_daily"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    # P0-A: Use BIGINT for autoincrement IDs (production scale)
+    id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
     tenant_id: Mapped[str] = mapped_column(TEXT, nullable=False, index=True)
     usage_date: Mapped[date] = mapped_column(DATE, nullable=False)
